Add login_timeout as a configurable parameter; bump the default to 30s
It used to be hardwired, to 10s.  I'd like to be able to specify it on
a per-connection basis.  And 10s does seem a bit slow for normal usage.

Update the tests too - however, we still want a short timeout, so set
it back to 10.

And test that it's set correctly.