Changes - add note about 1.00 tag created by Module::Release breakage
Also, reorder the entries, most recent first.

Module::Release managed to release the code, but since it has no
understanding of three-point version numbers, it mistakenly released
it as 1.00 instead of v0.1.0.

I've attempted to fix Module::Release such that it doesn't do this in
future, but it's gnarly code and BDFOY has not made a general release
that can cope with three point versions.