MyTest::Mock::ConnectParams - reinvoke mock-ssh using the same perl executable
Since mock-ssh is a Perl script which references /usr/bin/perl in the
shebang, this means it will always use that version of Perl if invoked
directly on shebang-respecting systems.  However, that may not exist,
or be of a version which doesn't play nicely with this script.

Therefore, invoke this script using the version of Perl running the
test code.  This bypasses the shebang and should help avoid problems
like that reported in RT bug #77013.