t/00.load.t - reinvoke mock-ssh using the same perl executable
This test compiles scripts using "perl -cw". However, that may not
exist in the path, or be of a version which doesn't play nicely with
this script.

Therefore, invoke this script using the version of Perl running the
test code.  This should help avoid problems like that reported in RT
bug #77013.